changed altituide from short to int

+++ vehicle-tracking-service/src/main/java/org/securityapps/vehicletracking/Netty/inbound/model/GpsMessage.java
@@ -7,7 +7,7 @@ public record GpsMessage(
     double latitude,
     double longitude,
     int speedKph,
-    short altitude,
+    int altitude,
     int angle,
     int satellites,
     int priority,
